fix: restore table of contents on pages that use the blog-post template
The ToC now shows up in the sidebar, thus has the advantage of always staying visible, at least on non-mobile screens.

This layout (toc inside the sidebar) also forces me to keep the headings concise so as to keep the sidebar from widening too much.

The min-width of the main pane of a with-sidebar had to be adjusted so that the layout doesnt switch to a top-bottom mobile layout too early.