fix ToC layout by replacing media query with a container query
In a `@media(min-width: 131ch)` rule, the value `131ch`, being a relative unit, is based on the *initial* value of `font`, not on the value set by user in stylesheet.

So, browser's calculations, which uses a different (smaller) font to calculate the aforementioned min-width, and my calculations, which uses (the larger) Atkinson font to calculate the min-width, fail to align.

As a result, ToC moves to a fixed layout way before it is expected to, cutting half of it off the screen.