Replace caddy with devd
After a while of running a caddy dev server,
it starts serving stale responses.
It does not refresh even after its restarted.

Why not miniserve?
miniserve is a file server more than a dev server.
So it does not auto-serve index.html,
which makes it useless for this project.

devd is aimed to be a dev server.