Publish a note - forest in plastic
I had to place an image related to the note in /static/images dir instead of putting both the note and the image in their own dir.

For notes, relative urls like <img src="./relative-link"> don't work because a note content is copied and displayed as-is on /notes page. But, the image's relative link breaks on /notes page.

Between the choice of maintaining a fragile script to re-write relative links just for the /notes page and placing images as their own entity, nix-style (each derivation has its own place in /nix/store – it is sym-linked to several places from there),  in /static/images, the latter was a much simpler and maintainable choice.